Test antiXSS middleware
