+++ server/routes/api.js
@@ -1,11 +1,10 @@
 const router = require("express").Router()
-const antiXSS = require("../private/js/utility")
 
 router.get("/api/test", (req, res) => {
     res.status(200).send("Working..");
 })
 
-router.post("/api/test",antiXSS, (req, res) => {
+router.post("/api/test", (req, res) => {
     res.status(200).json(req.body.hobbies[0].swimming);
 })
