+++ server/server.js
@@ -8,7 +8,7 @@ const bodyParser = require("body-parser")
 
 
 //Middleware imports
-//const antiXSS = require("./private/js/utility");
+const antiXSS = require("./private/middlewares/antiXSS");
 
 //Routes
 const apiRouter = require("./routes/api")
@@ -22,7 +22,7 @@ app.use(express.json({limit: "50mb"}))
 app.use(bodyParser.urlencoded({limit: '50mb', extended: true, parameterLimit: 50000}))
 
 // XSS protection
-//app.use(antiXSS)
+app.use(antiXSS)
 
 //Routes middleware
 app.use(apiRouter);
